refactor: improve error handling (#249)
* refactor: move `handleError` calls to interceptors

* refactor: only log errors with unknown code

* chore: explict check for sql.ErrNoRows when querying tenant

* fix: wrong type cast in tests

+++ rpcserver/rpcserver_test.go
@@ -1976,7 +1976,7 @@ func TestAutoSwap(t *testing.T) {
 				budget := status.Lightning.Budget
 				require.NoError(t, err)
 				require.Equal(t, 1, int(budget.Stats.Count))
-				require.Less(t, budget.Remaining, int64(budget.Total))
+				require.Less(t, budget.Remaining, budget.Total)
 				require.NotZero(t, budget.Stats.TotalFees)
 				require.NotZero(t, budget.Stats.TotalAmount)
 			})
